+++ backend/.env
@@ -1,2 +1,6 @@
 JWT_SECRET=my-32-characters-long-super-secure-secret-key-thats-from-stackoverflow
-REFRESH_TOKEN_SECRET=my-32-characters-long-super-secure-secret-key-thats-from-stackoverflow
+REFRESH_TOKEN_SECRET=my-32-characters-long-super-secure-refresh-secret-key-thats-from-the-internet
+MAILTRAP_HOST=sandbox.smtp.mailtrap.io
+MAILTRAP_PORT=2525
+MAILTRAP_USER=2ffdd7c9108046
+MAILTRAP_PASS=378385174926c2
