+++ backend/routes/database/database.js
@@ -1,0 +1,21 @@
+const mysql = require('mysql');
+
+// Create a MySQL connection
+const db = mysql.createConnection({
+    host: 'localhost',
+    user: 'root',
+    password: '',  // Add your password if needed
+    database: 'garrisons'
+});
+
+// Connect to the database
+db.connect((err) => {
+    if (err) {
+        console.error('Error connecting to MySQL:', err);
+        return;
+    }
+    console.log('Connected to MySQL');
+});
+
+// Export the connection for use in other files
+module.exports = db;
